requirements.txt
fastapi
uvicorn[standard]
pandas
python-dateutil
pydantic
python-dotenv
httpx
openpyxl
